jsystem: Fix hang when generating a lot of output
In case the process would write a lot of data to it's, stdout it would
at some point hang due to the output buffer being full.
Since jsystem would wait for the process to complete, both would be
stuck forever.

We now read the process's stdout before waiting for it's completion,
since reading is also a blocking operation.