bootstrap: Support running tools from any directory.
When searching for the WFDB, the script specifies an absolute search
path for wfdb tools. This means that we can run the script from any
matlab folder (not just the repo root), and then later when running the
toolbox functions the wfdb tools will still be found (since their
absolute path was saved).