default: Update power_methods
Set default to an empty list so that the first method is used by
default.
There was a bug that caused an error when the 'methods' parameter was
set to a value that didn't include the Lomb method because the
default 'power_methods' parameter contained the Lomb method by default.