Bugfix for records in PWD
When reading a record from the current matlab PWD an error would occur.
This happens because the built in matlab 'fileparts' function returns an
empty string as the path in such a case.
Wrapped this function in a custom method that simply handles this case
and returns '.' (the PWD) as the path.

The wrapper has a different name than the builtin because we don't want
to break code that uses this library in case it depends on the original
builtin 'fileparts' and expects this behavior.