Bugfix with Scanner delimiter pattern
The Scanner pattern used was supposed to be '\A' which matches the
beginning of input (see
https://docs.oracle.com/javase/8/docs/api/java/util/regex/Pattern.html).
In java strings the backslash must be escaped, so to enter '\A' as a
delimiter the string "\\A" must be passed into the scanner/Patter
constructor. That would work in java code.

However since this is matlab, turns out the escaping is not required...
The bug was that since '\\A' was passed from matlab, the java Pattern
class interpreted it as a literal backslash '\' followed by an 'A' which
is not what we want.